feat: first pass at mid-agent
I can't keep using Ansible as an execution engine forever. It is great
for rapid prototyping but it SUCKS hardcore to integrate and debug. I
was always planning on eventually replacing it but it seems the time has
come a bit sooner than I realized. This is very much not production
ready (nor is the rest of the thing) but it's a start.