refactor: switch to single Ansible bundle and just installing it once
This is _way_ more efficient overall for multiple reasons. The first is
that we aren't storing tons of duplicate data for each module (which
always includes a good chunk of `module_utils`) so the Go binary size is
reduced. We also don't have to transmit the contents over the wire for
every RPC. We could probably do some caching stuff on the remote end but
this is easier. The bundle size is honestly not that big. Also this uses
Gzip compression instead of ZIP so it's a bit more space efficient. The
one downside is that it stays resident on the remote machine and doesn't
clean up after itself, but the agent doesn't do that anyway so it's fine
for now.