feat: add support for more `community.general` modules
Using some basic heuristics I pulled in any modules that might be useful
for single-machine management. Anything that is made for managing
external or cloud services (e.g. `keycloak_`, `proxmox_`, etc) was not
included. This was not a perfect filter and some things that should have
been skipped might have gotten through as well as the inverse.