refactor!: rewrite File resource
it's still not able to handle updating on refresh but it's more feasible
to implement that now I think.

i've been wanting to rewrite this for a while now, ever since the agent
was introduced, because this resource relied so heavily on Ansible
Playbook behavior. At some point I started a rewrite and abandoned it,
then half-rewrote parts of it to hack in agent support. It sucked, but
it proved that it was achievable. This is finally in a much better spot
under the hood despite still having some of the same issues (and some
new ones) as the last implementation. The only breaking API change is
that `ensure=touch` has been removed. It was redundant; `ensure=file`
should just work the way you expect.