feat: AnsibleTaskList
This is a hack. A backdoor. An escape hatch. Avoid using it if you can,
but it's here if you _really_ need it.

I will probably end up removing it at some point but there's just too
many missing "first-class" resources that I think this has to exist
otherwise this project is not really usable.