refactor!: take connection and options as resource arguments
I keep running into issues where having to set up the connection in the
provider is causing a lot of issues. I think I need to be able to
set the connection and any config args per-resource while doing a
deep-merge with any options specified in the main provider config (for
backwards compatibility and ergonomics where it makes sense).