feat: add ability to disable dry run checks
Normally these checks are good for catching any issues that make it past
the static analysis, but sometimes it returns an error when it
shouldn't or it just takes way too long. This is an "escape hatch" to
disable that functionality. Right now it is default enabled but might be
default disabled in the future potentially.